+++ glide/core/mean_inference_result/__init__.py
@@ -1,9 +1,0 @@
-from glide.core.mean_inference_result.base import MeanInferenceResult
-from glide.core.mean_inference_result.classical import ClassicalMeanInferenceResult
-from glide.core.mean_inference_result.prediction_powered import PredictionPoweredMeanInferenceResult
-
-__all__ = [
-    "MeanInferenceResult",
-    "ClassicalMeanInferenceResult",
-    "PredictionPoweredMeanInferenceResult",
-]